Factor out SearchTestCase for usage elsewhere
